+++ polaris-react/package.json
@@ -51,8 +51,8 @@
     "test": "jest",
     "clean": "rm -rf .turbo node_modules build build-internal",
     "start": "serve ./build-internal/storybook/static -l ${PORT:=6006}",
-    "storybook": "start-storybook -p 6006 --quiet",
-    "storybook:build": "build-storybook -o build-internal/storybook/static",
+    "storybook": "storybook dev -p 6006 --quiet",
+    "storybook:build": "storybook build -o build-internal/storybook/static",
     "preversion": "node ./scripts/readme-update-version"
   },
   "dependencies": {
@@ -74,18 +74,25 @@
     "@shopify/jest-dom-mocks": "^3.0.5",
     "@shopify/postcss-plugin": "^5.0.1",
     "@shopify/react-testing": "^4.1.0",
-    "@shopify/storybook-a11y-test": "1.0.1",
-    "@storybook/addon-a11y": "^6.5.12",
-    "@storybook/addon-toolbars": "^6.5.12",
-    "@storybook/addon-viewport": "^6.5.12",
-    "@storybook/builder-webpack5": "^6.5.12",
-    "@storybook/manager-webpack5": "^6.5.12",
-    "@storybook/react": "^6.5.12",
+    "@shopify/storybook-a11y-test": "1.1.3",
+    "@storybook/addon-a11y": "^7.6.4",
+    "@storybook/addon-toolbars": "^7.6.4",
+    "@storybook/addon-viewport": "^7.6.4",
+    "@storybook/api": "^7.6.4",
+    "@storybook/blocks": "^7.6.4",
+    "@storybook/components": "^7.6.4",
+    "@storybook/manager-api": "^7.6.4",
+    "@storybook/preview-api": "^7.6.4",
+    "@storybook/react": "^7.6.4",
+    "@storybook/react-vite": "^7.6.4",
+    "@types/node": "^16.11.11",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
+    "@types/react-transition-group": "^4.4.2",
+    "babel-core": "7.0.0-bridge.0",
+    "babel-loader": "^8.1.0",
     "change-case": "^3.1.0",
     "chromatic": "^6.5.4",
-    "create-file-webpack": "^1.0.2",
     "globby": "^11.1.0",
     "js-yaml": "^4.1.0",
     "node-sass": "^9.0.0",
@@ -99,7 +106,10 @@
     "react-dom": "^18.2.0",
     "rimraf": "^3.0.0",
     "sass-loader": "^12.4.0",
-    "webpack": "5"
+    "serve": "^14.2.0",
+    "storybook": "^7.6.4",
+    "vite": "^4.4.9",
+    "vite-plugin-replace": "^0.1.1"
   },
   "browserslist": [
     "last 3 chrome versions",
